+++ backend/tarsy/agents/iteration_controllers/react_iteration_controller.py
@@ -16,7 +16,7 @@ that LLMs are specifically trained to handle.
 from typing import TYPE_CHECKING
 
 from tarsy.utils.logger import get_module_logger
-from tarsy.models.llm import LLMMessage
+from tarsy.models.unified_interactions import LLMMessage
 from .base_iteration_controller import IterationController, IterationContext
 
 if TYPE_CHECKING:
