+++ backend/tarsy/agents/iteration_controllers/react_tools_partial_controller.py
@@ -8,7 +8,7 @@ stage-specific analysis, providing incremental insights during chain processing.
 from typing import TYPE_CHECKING
 
 from tarsy.utils.logger import get_module_logger
-from tarsy.models.llm import LLMMessage
+from tarsy.models.unified_interactions import LLMMessage
 from .base_iteration_controller import IterationController, IterationContext
 
 if TYPE_CHECKING:
