+++ backend/tarsy/hooks/typed_history_hooks.py
@@ -12,7 +12,7 @@ from typing import Any, Dict, Optional
 
 from tarsy.hooks.typed_context import BaseTypedHook
 from tarsy.models.unified_interactions import LLMInteraction, MCPInteraction
-from tarsy.models.history import StageExecution
+from tarsy.models.db_models import StageExecution
 from tarsy.services.history_service import HistoryService
 
 logger = logging.getLogger(__name__)
