+++ backend/tarsy/models/__init__.py
@@ -1,5 +1,5 @@
-# Models package
-from .llm import LLMMessage
-from .chains import ChainDefinitionModel, ChainStageModel
+# Models package - Phase 5: Minimal exports to avoid circular imports
+# Configuration models should be imported directly from agent_config.py to avoid circular imports
+from .unified_interactions import LLMMessage
 
-__all__ = ["LLMMessage", "ChainDefinitionModel", "ChainStageModel"] 
+__all__ = ["LLMMessage"]
