+++ backend/tests/unit/services/test_alert_service_template_integration.py
@@ -15,7 +15,7 @@ from unittest.mock import patch, Mock, AsyncMock
 from tarsy.config.settings import Settings
 from tarsy.services.alert_service import AlertService
 from tarsy.models.alert_processing import AlertProcessingData
-from tarsy.agents.constants import IterationStrategy
+from tarsy.models.constants import IterationStrategy
 
 
 @pytest.mark.unit
